app/requirements.txt
aiofiles==24.1.0
aiohttp==3.11.11
aiosqlite==0.21.0
boto3==1.35.92
browserforge==1.2.3
fake_useragent==2.0.3
fastapi==0.115.8
google_api_python_client==2.157.0
google_auth_oauthlib==1.2.1
httpx==0.28.1
imap_tools==1.8.0
ipython==8.12.3
loguru==0.7.3
names==0.3.0
nest_asyncio==1.6.0
numpy==2.2.3
orjson==3.10.15
pandas==2.2.3
protobuf==5.29.3
pydantic==2.10.6
pyotp==2.9.0
python-dotenv==1.0.1
redis==5.2.1
Requests==2.32.3
rq==2.1.0
scipy==1.15.1
thefuzz==0.22.1
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
websockets==14.2
zendriver==0.4.3
rq-scheduler
